requirements.txt
flask
nomic.gpt4all